+++ api/account/src/effect/config.rs
@@ -1,7 +1,7 @@
 #[cfg_attr(test, mockall::automock)]
 pub trait Config {
     fn firebase_project_id(&self) -> &str;
-    fn max_connections(&self) -> &usize;
+    fn max_connections(&self) -> &u32;
 }
 
 #[cfg_attr(test, mockall::automock(type Config = MockConfig;))]
